feat(money): store and split bills at 2-decimal precision
Money is now real 2-decimal som instead of whole-som integers: columns
become numeric(14,2) (Drizzle mode:'number'), so the DB holds the exact
value (e.g. 33333.33) and code handles it as a plain number — no scaling.

Splitting truncates each share to the cent via to2() (a dust-guarded
floor) and keeps equal sharers equal; the bill total holds the true
aggregate, so the small leftover is carried by total rather than nudged
onto a participant (sum(shares) ≤ total, gap sub-som). round100 and the
largest-payer reconciliation are gone. Decimal item prices (e.g. 10.33)
are now accepted in both the Mini App and the bot.

Migration 0001 is an in-place ALTER TYPE (value-preserving: 50000 →
50000.00, no ×100); it also adds the previously untracked
contacts.linked_telegram_id column idempotently.

+++ src/server/schemas.ts
@@ -1,13 +1,13 @@
 // Zod request schemas for the Mini App API. Pure (no DB) — unit-tested.
-// Money fields arrive as JS numbers (whole som); the mappers convert to bigint.
+// Money fields arrive as JS numbers (real 2-decimal som); stored as numeric(14,2).
 
 import { z } from 'zod'
 
 export const createBillItemSchema = z.object({
   name: z.string().trim().max(200).default(''),
-  // Whole som, must be a positive integer (empty placeholder rows are filtered
-  // client-side before submit).
-  price: z.number().int().positive(),
+  // Positive som; decimals allowed (e.g. 10.33). Empty placeholder rows are
+  // filtered client-side before submit.
+  price: z.number().positive(),
   shareContactIds: z.array(z.uuid()).min(1),
 })
 
@@ -17,7 +17,7 @@ export const createBillSchema = z
     participantContactIds: z.array(z.uuid()).min(1),
     items: z.array(createBillItemSchema).min(1),
     servicePct: z.number().min(0).max(100).default(0),
-    tip: z.number().int().nonnegative().default(0),
+    tip: z.number().nonnegative().default(0),
   })
   .superRefine((data, ctx) => {
     // Every item's sharers must be among the bill's participants.
